chore: combine passport info functions
