--- sample/Assets/Plugins/AltTester.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3f621702f1ccd4d5fb7009af90cf285e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
