--- sample/Assets/Plugins/AltTester/AltTesterEditorSettings.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 287fa8cf2344747edb0a62da78598176
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
