--- sample/Assets/Scripts/Passport/DeviceCodeLogin/DeviceCodeLoginScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d981605e6a89d41e69a431e67f770872
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
