--- sample/Assets/Scripts/Passport/GetAccessToken/GetAccessTokenScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7d86f111564dd4f44886a841bc567be3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
