--- sample/Assets/Scripts/Passport/GetIdToken/GetIdTokenScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c9e0a2073cd814a19b3eef985d78b1f9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
