--- sample/Assets/Scripts/Passport/GetLinkedAddresses/GetLinkedAddressesScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3dde6b956c3374cc58342446b062c29b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
