+++ sample/Assets/Scripts/Passport/GetUserInfo/GetUserInfoScript.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 using Cysharp.Threading.Tasks;
@@ -56,6 +57,87 @@ public class GetUserInfoScript : MonoBehaviour
         catch (System.Exception ex)
         {
             ShowOutput($"Failed to get Passport ID: {ex.Message}");
+        }
+    }
+
+    /// <summary>
+    /// Retrieves the currently logged-in user's access token.
+    /// </summary>
+    public void GetAccessToken()
+    {
+        GetAccessTokenAsync();
+    }
+
+    private async UniTaskVoid GetAccessTokenAsync()
+    {
+        if (Passport.Instance == null)
+        {
+            ShowOutput("Passport instance is null");
+            return;
+        }
+        try
+        {
+            string accessToken = await Passport.Instance.GetAccessToken();
+            ShowOutput(accessToken);
+        }
+        catch (System.Exception ex)
+        {
+            ShowOutput($"Failed to get access token: {ex.Message}");
+        }
+    }
+
+    /// <summary>
+    /// Retrieves the currently logged-in user's ID token.
+    /// </summary>
+    public void GetIdToken()
+    {
+        GetIdTokenAsync();
+    }
+
+    private async UniTaskVoid GetIdTokenAsync()
+    {
+        if (Passport.Instance == null)
+        {
+            ShowOutput("Passport instance is null");
+            return;
+        }
+        try
+        {
+            string idToken = await Passport.Instance.GetIdToken();
+            ShowOutput(idToken);
+        }
+        catch (System.Exception ex)
+        {
+            ShowOutput($"Failed to get ID token: {ex.Message}");
+        }
+    }
+
+    /// <summary>
+    /// Retrieves the user's linked external wallets from the Passport account dashboard.
+    /// </summary>
+    public void GetLinkedAddresses()
+    {
+        Debug.Log("E2E TEST DEBUG: GetLinkedAddresses() called (public method)");
+        GetLinkedAddressesAsync();
+    }
+
+    private async UniTaskVoid GetLinkedAddressesAsync()
+    {
+        Debug.Log("E2E TEST DEBUG: GetLinkedAddressesAsync() started");
+        if (Passport.Instance == null)
+        {
+            ShowOutput("Passport instance is null");
+            return;
+        }
+        try
+        {
+            List<string> addresses = await Passport.Instance.GetLinkedAddresses();
+            string outputMessage = addresses.Count > 0 ? string.Join(", ", addresses) : "No linked addresses";
+            ShowOutput(outputMessage);
+        }
+        catch (System.Exception ex)
+        {
+            ShowOutput($"Failed to get linked addresses: {ex.Message}");
         }
     }
