+++ sample/Assets/Scripts/Passport/PassportInitialisation.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: b513e9e94967f45dbbf2c4e1f36dd77e
-PrefabImporter:
+guid: 2ec54a43e70e741e5812e8e6f78101a8
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
