+++ sample/Assets/Scripts/Passport/PassportInitialisation/PassportInitializationScript.cs
