--- sample/Assets/Scripts/Passport/PkceAuth/PkceAuthScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f0df226c65f9c4e8e940ad3333e5f8e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
